timebank functionality

+++ kcommons/src/main/java/com/kairos/dto/activity/time_bank/CTARuletemplateBonus.java
@@ -16,5 +16,5 @@ import java.util.List;
 public class CTARuletemplateBonus  implements Serializable {
 
     private List<CTADistributionDTO> ctaDistributions;
-    private long ctaBonusMinutes;
+    private double ctaBonusMinutes;
 }
