+++ kcommons/src/main/java/com/kairos/dto/activity/time_bank/time_bank_basic/time_bank/CTADistributionDTO.java
@@ -20,7 +20,7 @@ import java.math.BigInteger;
 public class CTADistributionDTO  implements Serializable {
     private BigInteger id;
     private String name;
-    private int minutes;
+    private double minutes;
     private float cost;
 
 }
